+++ test/String.hs
@@ -12,7 +12,7 @@ import           Test.Hspec as H
 import qualified Streamly.Memory.Array as A
 import qualified Streamly.Internal.Memory.ArrayStream as AS
 import qualified Streamly.Prelude as S
-import qualified Streamly.Data.Unicode.Stream as SS
+import qualified Streamly.Unicode.Stream as SS
 import qualified Streamly.Internal.Unicode.Stream as IUS
 import qualified Streamly.Internal.Memory.Unicode.Array as IUA