+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### Breaking changes
 
+* Deprecate `Streamly.Data.Unicode.Stream` in favor of `Streamly.Unicode.Stream`
 * Change the signature of `foldlM'` to make the initial value of the
   accumulator monadic.
 * Change the signature of `concatMapWith` to ensure that it can be